feat(logger): move logger to own repository
BREAKING CHANGE: This module no longer exports the 'logger'. It was moved
                 to 'openwhisk-action-logger'

fixes #42
