+++ package.json
@@ -18,9 +18,8 @@
     "semantic-release": "semantic-release"
   },
   "dependencies": {
-    "@adobe/helix-log": "4.1.0",
-    "bunyan": "^1.8.12",
-    "dotenv": "^8.2.0",
+    "@adobe/helix-log": "^4.1.0",
+    "bunyan": "1.8.12",
     "openwhisk": "^3.20.0",
     "serverless-http": "^2.3.0"
   },
@@ -38,7 +37,6 @@
     "ghooks": "2.0.4",
     "jsdoc-to-markdown": "5.0.3",
     "lint-staged": "9.4.3",
-    "lodash": "4.17.15",
     "mocha": "6.2.2",
     "mocha-junit-reporter": "1.23.1",
     "nock": "11.7.0",
